#915 refactor: extract product detail
