+++ src/features/product/catalog/components/badges/PriceBadge.tsx
@@ -1,10 +1,11 @@
 import { Badge } from "@/components/ui/badge.tsx";
 import { cn } from "@/lib/utils.ts";
 import { TrendingDown, TrendingUp, Tag, XCircle } from "lucide-react";
-import type { PriceEventType } from "@/types/events.ts";
 import { useTranslation } from "react-i18next";
 import type { TFunction } from "i18next";
 import { useMemo } from "react";
+
+type PriceEventType = "PRICE_DISCOVERED" | "PRICE_DROPPED" | "PRICE_INCREASED" | "PRICE_REMOVED";
 
 interface PriceBadgeProps {
     readonly eventType: PriceEventType;
