+++ src/features/product/detail/api/useSimilarProducts.ts
