+++ src/features/product/detail/components/ProductDetailPageSkeleton.tsx
