+++ src/features/product/detail/components/ProductImageGallery.tsx
