+++ src/features/product/detail/components/ProductSectionHeading.tsx
