+++ src/features/product/detail/components/dealer/ProductDealerItems.tsx
@@ -1,8 +1,8 @@
-import { useDealerProducts } from "@/hooks/shop/useDealerProducts.ts";
+import { useDealerProducts } from "@/features/product/detail/api/useDealerProducts.ts";
 import { ProductGridItem } from "@/features/product/catalog/components/grid/ProductGridItem.tsx";
 import { ProductGridItemSkeleton } from "@/features/product/catalog/components/grid/ProductGridItemSkeleton.tsx";
-import { ProductSectionHeading } from "@/components/product/detail/ProductSectionHeading.tsx";
-import { ProductCarouselNavButtons } from "@/components/product/detail/ProductCarouselNavButtons.tsx";
+import { ProductSectionHeading } from "@/features/product/detail/components/ProductSectionHeading.tsx";
+import { ProductCarouselNavButtons } from "@/features/product/detail/components/ProductCarouselNavButtons.tsx";
 import { EmptyState } from "@/components/common/EmptyState.tsx";
 import { useTranslation } from "react-i18next";
 import { ArrowUpRight, ServerCrash } from "lucide-react";
