+++ src/features/product/detail/lib/events/types.ts
