+++ src/features/product/detail/lib/seo/productHeadMeta.ts
@@ -1,5 +1,5 @@
 import type { PersonalizedGetProductData } from "@/client";
-import { generateProductJsonLdScript } from "@/lib/seo/productJsonLd.ts";
+import { generateProductJsonLdScript } from "./productJsonLd.ts";
 import { BANNER_IMAGE_URL } from "@/lib/seo/seoConstants.ts";
 import { env } from "@/env.ts";
 import { generateHreflangLinks } from "@/lib/seo/hreflangLinks.ts";
